Resolved the problem with removing measures and when opening a inv3 file and all measures are showed in the same slice
The problem with removing measures was because a change in pubsub: it
only accepts strings in the sendMessage message parameter or tuples of
string. I was using int. So it was only necessary to change that in this
way: "Remove Actor " + str(slice_number). The problem when opening an
inv3 file was because I was not verifying if the measure being added is
from the actual showed slice.

Squashed commit of the following:

commit 59bd23b0141aaa504b4af5a3cdc781b07e09889f
Author: Thiago Franco de Moraes <totonixsame@gmail.com>
Date:   Mon Jun 25 14:43:53 2012 -0300

    When opening a inv3 file only the measures from the showing slice in showed

commit 07fc45a907438cb7ebdcaca84d47f2b04d301da2
Author: Thiago Franco de Moraes <totonixsame@gmail.com>
Date:   Mon Jun 25 14:24:47 2012 -0300

    Adding measures actor from saved inv3 when it's opened

commit 46b02de95bd2f1676a8bd96b18c48301047f9939
Author: Thiago Franco de Moraes <totonixsame@gmail.com>
Date:   Mon Jun 25 14:17:15 2012 -0300

    Removing the measures is working again