Using replace method error to decode dicom fields and avoid crashes.
Some dicom files have the encoding setted wrongly, so when decoding a
dicom field from string to unicode some errors may happen. Python string.decode
have some methods to handle those cases. We are using the "replace" one,
the characters it cannot decode to unicode it replaces with a character
showing the character was not decoded.