Adds boolean operation to mask (union, xor, diff, intersection)
Implements boolean operations to masks. The user has a gui to choose the
masks and the operation (union, xor, diff, intersection). Then
InVesalius creates a new mask with the result of the operation.

Doing some boolean operations

Added Intesection and XOR

There was an error in the bool diff operation

Gui improvements

changed "xor" to "exclusive disjuction"

added the icons

Improvements to the boolean operation dialog

Added a menu to mask operations (booleans is there)