Encoding Errors in Windows (#110)
Fix the above unicode related errors:
- InVesalius was not starting in Windows with users with non ascii chars.
- Not opening dicom files from path with non ascii chars.
- Not opening bitmap files from path with non ascii chars.
- Not opening Analyze, Nifti with filename with non ascii chars.
- Not exporting surface when the path has non ascii chars.
- Not exporting picture when the path has non ascii chars.
- Not importing surface when the path has non ascii chars.
- Strange chars when showing developers and translators with non ascii chars.

To fix that errors we are convert all paths and filenames to unicode. wxPythons eases this, since it gives the paths and filenames as unicode. When using VTK to save or open a file we encode the filename to sys.getfilesystemencoding. In Windows it is necessary an additional step, convert the filename to an short representation using [win32api](http://docs.activestate.com/activepython/2.6/pywin32/win32api__GetShortPathName_meth.html).