Fix xyspacing in invesalius/control.py
Running Invesalius via CLI to export DICOM/Bitmap to 3D (STL, OBJ et all) generated stretched 3Ds while in GUI everything was fine. The bug was present in OpenBitmapFiles() and OpenDicomGroup(), I'll refer to the latter method: after calculating resolution_percentage (line 800) when in GUI xyspacing was adjusted accordingly (line 814) but, since via CLI the if condition (line 802) it's not met, this caused discrepancies in the 3d dimensions. 

This subtle bug was probably caused due to an error in indentation, moving line 814 outside of line 802 if, fixed the issue in my scenarios and should be sufficient.