FIX: Using pedal for setting fiducials
Previously, the callback for setting the fiducial was removed while
executing the callback. However, that did not work because it made
the list of callbacks that PedalConnection looped over change on the
fly. Instead, use 'remove when released' option for callbacks to
automatically remove the callback when the pedal is released.