MOD: Trigger pulse by sending BREAK via serial port
Previously, when using high baud rates, the pulse signal was not
long enough to trigger the TMS device.

Instead of sending a byte via serial port (for which the duration of
the signal is determined implicitly by the baud rate), trigger the
pulse by sending BREAK signal, for which the pulse duration can be
explicitly defined.

Set the pulse duration tentatively to 5 ms, adjust later if needed.