ADD: Checkbox to allow stimulating when coil is not at target
Add a "Lock to target" checkbox to the UI. When checked, only allow
triggering stimulation pulse if the coil is at the target. When unchecked,
allow triggering stimulation pulse regardless of the coil location.

Keep the checkbox disabled until a target is selected. When a target
is selected, enable and automatically check the checkbox so that the
default mode of operation is to allow triggering stimulation pulse
only when the coil is at the target.