FIX: Incorrect voxel to world space transformation function
The function is called to add world coordinates for markers to
be stored as csv columns when the markers are saved on the disk.

The previous implementation assumed that the markers, as stored by
MarkersPanel class, are in a coordinate system with reverted y-axis,
and, hence, called an auxiliary function that reverted the y-coordinate.

However, MarkersPanel class stores the marker coordinates based on the
value received by this subscriber (in MarkersPanel class):

Publisher.subscribe(self.UpdateCurrentCoord, 'Set cross focal point')

All the publishers of 'Set cross focal point' do the inversion of
y-coordinate if they operate in space with inverted y-coordinate,
see, e.g., this line in styles_3d.py, picking points from the
volume viewer:

Publisher.sendMessage('Set cross focal point', position=[x, -y, z, None, None, None])

Or, alternatively, they operate already in voxel space, such as
this publisher in styles.py, picking points from the slice viewers:

Publisher.sendMessage('Set cross focal point', position=[x, y, z, None, None, None])

In addition, I checked that picking a marker from anterior part
of the head yields an input argument that has a higher y-coordinate
than picking a marker from posterior part of the brain - implying
that the assumption that the y-coordinate is inverted, as compared
to the voxel space (where y-axis grows into anterior direction)
when the function is called is incorrect.